+++ tests/_data/infiles/environment/broken-env/README.md
@@ -1,0 +1,2 @@
+A dir that MUST NOT contain any (python (virtual) environment.
+This is used in integration tests.
