+++ tests/_data/infiles/environment/broken-env/broken-json.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python3
+print(r'{"some unfinished json')
+exit(0)
