+++ tests/_data/infiles/environment/broken-env/non-zero.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python3
+print(r'[]')
+exit(1337)
